+++ src/cpu/tests/call.rs
@@ -5,16 +5,16 @@ fn test_call() {
     let (mut cpu, mut ram) = init(None);
     cpu.reset();
 
-    ram[cpu.pc as usize] = 0xCD;
-    ram[cpu.pc as usize + 1] = 0x22;
-    ram[cpu.pc as usize + 2] = 0x11;
+    ram[cpu.pc] = 0xCD;
+    ram[cpu.pc + 1] = 0x22;
+    ram[cpu.pc + 2] = 0x11;
 
     let old_cycles = cpu.cycles;
     cpu.step(&mut ram);
     assert_eq!(cpu.cycles, old_cycles + 24);
     assert_eq!(cpu.pc, 0x1122);
-    assert_eq!(ram[cpu.sp as usize + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
-    assert_eq!(ram[cpu.sp as usize + 2], (0x103_u16 & 0xFF) as u8);
+    assert_eq!(ram[cpu.sp + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
+    assert_eq!(ram[cpu.sp + 2], (0x103_u16 & 0xFF) as u8);
 }
 
 #[test]
@@ -22,9 +22,9 @@ fn test_call_nz() {
     let (mut cpu, mut ram) = init(None);
     cpu.reset();
 
-    ram[cpu.pc as usize] = 0xC4;
-    ram[cpu.pc as usize + 1] = 0x22;
-    ram[cpu.pc as usize + 2] = 0x11;
+    ram[cpu.pc] = 0xC4;
+    ram[cpu.pc + 1] = 0x22;
+    ram[cpu.pc + 2] = 0x11;
 
     cpu.f.set_z();
 
@@ -46,8 +46,8 @@ fn test_call_nz() {
     assert_eq!(cpu.cycles, old_cycles + 24);
     assert_eq!(cpu.pc, 0x1122);
     assert_ne!(cpu.sp, 0xFFFE);
-    assert_eq!(ram[cpu.sp as usize + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
-    assert_eq!(ram[cpu.sp as usize + 2], (0x103_u16 & 0xFF) as u8);
+    assert_eq!(ram[cpu.sp + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
+    assert_eq!(ram[cpu.sp + 2], (0x103_u16 & 0xFF) as u8);
 }
 
 #[test]
@@ -55,9 +55,9 @@ fn test_call_z() {
     let (mut cpu, mut ram) = init(None);
     cpu.reset();
 
-    ram[cpu.pc as usize] = 0xCC;
-    ram[cpu.pc as usize + 1] = 0x22;
-    ram[cpu.pc as usize + 2] = 0x11;
+    ram[cpu.pc] = 0xCC;
+    ram[cpu.pc + 1] = 0x22;
+    ram[cpu.pc + 2] = 0x11;
 
     cpu.f.unset_z();
 
@@ -79,8 +79,8 @@ fn test_call_z() {
     assert_eq!(cpu.cycles, old_cycles + 24);
     assert_eq!(cpu.pc, 0x1122);
     assert_ne!(cpu.sp, 0xFFFE);
-    assert_eq!(ram[cpu.sp as usize + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
-    assert_eq!(ram[cpu.sp as usize + 2], (0x103_u16 & 0xFF) as u8);
+    assert_eq!(ram[cpu.sp + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
+    assert_eq!(ram[cpu.sp + 2], (0x103_u16 & 0xFF) as u8);
 }
 
 #[test]
@@ -88,9 +88,9 @@ fn test_call_nc() {
     let (mut cpu, mut ram) = init(None);
     cpu.reset();
 
-    ram[cpu.pc as usize] = 0xD4;
-    ram[cpu.pc as usize + 1] = 0x22;
-    ram[cpu.pc as usize + 2] = 0x11;
+    ram[cpu.pc] = 0xD4;
+    ram[cpu.pc + 1] = 0x22;
+    ram[cpu.pc + 2] = 0x11;
 
     cpu.f.set_c();
 
@@ -112,8 +112,8 @@ fn test_call_nc() {
     assert_eq!(cpu.cycles, old_cycles + 24);
     assert_eq!(cpu.pc, 0x1122);
     assert_ne!(cpu.sp, 0xFFFE);
-    assert_eq!(ram[cpu.sp as usize + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
-    assert_eq!(ram[cpu.sp as usize + 2], (0x103_u16 & 0xFF) as u8);
+    assert_eq!(ram[cpu.sp + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
+    assert_eq!(ram[cpu.sp + 2], (0x103_u16 & 0xFF) as u8);
 }
 
 #[test]
@@ -121,9 +121,9 @@ fn test_call_c() {
     let (mut cpu, mut ram) = init(None);
     cpu.reset();
 
-    ram[cpu.pc as usize] = 0xDC;
-    ram[cpu.pc as usize + 1] = 0x22;
-    ram[cpu.pc as usize + 2] = 0x11;
+    ram[cpu.pc] = 0xDC;
+    ram[cpu.pc + 1] = 0x22;
+    ram[cpu.pc + 2] = 0x11;
 
     cpu.f.unset_c();
 
@@ -145,6 +145,6 @@ fn test_call_c() {
     assert_eq!(cpu.cycles, old_cycles + 24);
     assert_eq!(cpu.pc, 0x1122);
     assert_ne!(cpu.sp, 0xFFFE);
-    assert_eq!(ram[cpu.sp as usize + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
-    assert_eq!(ram[cpu.sp as usize + 2], (0x103_u16 & 0xFF) as u8);
+    assert_eq!(ram[cpu.sp + 1], ((0x103_u16 & 0xFF00) >> 8) as u8);
+    assert_eq!(ram[cpu.sp + 2], (0x103_u16 & 0xFF) as u8);
 }
